big roy's comment- code clean up
